update roles in fire store

+++ lib/features/oreder_details/domain/use_cases/update_order_state_usecase.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:injectable/injectable.dart';
 import 'package:tracking_app/config/base/base_response.dart';
 import 'package:tracking_app/core/storage/secure_storage_service.dart';
@@ -18,16 +19,20 @@ class UpdateOrderStateUseCase {
 
     if (response is SuccessBaseResponse<UpdateOrderStateResponseEntity>) {
       final driverId = await SecureStorageService.getDriverId();
-      print("Driver ID => $driverId");
-      if (driverId != null) {
-        if (params.state == "inProgress") {
-          await repo.saveCurrentOrder(
-            driverId: driverId,
-            orderId: params.orderId,
-            state: params.state,
-          );
-        } else if (params.state == "canceled") {
-          await repo.deleteCurrentOrder(driverId: driverId);
+
+      if (driverId != null && driverId.isNotEmpty) {
+        try {
+          if (params.state == "inProgress") {
+            await repo.saveCurrentOrder(
+              driverId: driverId,
+              orderId: params.orderId,
+              state: params.state,
+            );
+          } else if (params.state == "canceled") {
+            await repo.deleteCurrentOrder(driverId: driverId);
+          }
+        } catch (error) {
+          debugPrint('Current order sync failed: $error');
         }
       }
     }
